scripts/requirements.txt
instaloader==4.15
supabase==2.11.0
python-dotenv==1.0.1
requests==2.32.3
